Renaming Auto mapper Folder and Sync wrong name spaces
