+++ Garage3.0/Garage3.0.Web/Automapper/GarageMappings.cs
@@ -2,13 +2,13 @@
 using Garage3._0.Web.Models.Entities;
 using Garage3._0.Web.Models.ViewModels;
 
-namespace Garage3._0.Web.Garagemapper;
+namespace Garage3._0.Web.Automapper;
 
 public class GarageMappings : Profile
 {
     public GarageMappings()
     {
-        CreateMap<MemberEntity, MemberCreateViewModel>().ReverseMap(); 
+        CreateMap<MemberEntity, MemberCreateViewModel>().ReverseMap();
         CreateMap<VehicleEntity, VehicleIndexViewModel>();
 
     }
